+++ src/collectors/mod.rs
@@ -1,3 +1,4 @@
+pub mod diskstats;
 pub mod iostat;
 pub mod raid;
 pub mod smart;